+++ src/Resources/contao/dca/tl_content.php
@@ -16,10 +16,10 @@
 
 declare(strict_types=1);
 
-use Hofff\Contao\RateIt\EventListener\Dca\ContentBaseDcaListener;
+use Hofff\Contao\RateIt\EventListener\Dca\ContentDcaListener;
 
-$GLOBALS['TL_DCA']['tl_content']['config']['onsubmit_callback'][] = [ContentBaseDcaListener::class, 'insert'];
-$GLOBALS['TL_DCA']['tl_content']['config']['ondelete_callback'][] = [ContentBaseDcaListener::class, 'delete'];
+$GLOBALS['TL_DCA']['tl_content']['config']['onsubmit_callback'][] = [ContentDcaListener::class, 'insert'];
+$GLOBALS['TL_DCA']['tl_content']['config']['ondelete_callback'][] = [ContentDcaListener::class, 'delete'];
 
 /**
  * palettes
