+++ src/Resources/contao/dca/tl_settings.php
@@ -16,7 +16,7 @@
 
 declare(strict_types=1);
 
-use Hofff\Contao\RateIt\EventListener\Dca\SettingsBaseDcaListener;
+use Hofff\Contao\RateIt\EventListener\Dca\SettingsDcaListener;
 
 /**
  * palettes
@@ -74,7 +74,7 @@ $GLOBALS['TL_DCA']['tl_settings']['fields']['rating_template'] = [
     'default'          => 'rateit_default',
     'exclude'          => true,
     'inputType'        => 'select',
-    'options_callback' => [SettingsBaseDcaListener::class, 'getRateItTemplates'],
+    'options_callback' => [SettingsDcaListener::class, 'getRateItTemplates'],
     'eval'             => ['mandatory' => true, 'tl_class' => 'w50'],
 ];
